% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc_glmnet_sumstats}
\alias{auc_glmnet_sumstats}
\title{calculate AUC from glmnet_sumstats results}
\usage{
auc_glmnet_sumstats(r2, ncase, ncont)
}
\arguments{
\item{r2}{model R2}

\item{ncase}{number of cases}

\item{ncont}{number of controls}
}
\value{
AUC
}
\description{
calculate AUC from glmnet_sumstats results
}
