% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sumstats.R
\name{combine_sumstats}
\alias{combine_sumstats}
\title{Combine summary statistics}
\usage{
combine_sumstats(sumstats)
}
\arguments{
\item{sumstats}{list of sumstats objects}
}
\value{
sumstats object that results from summing over the elements of the input list, with additional attributes 'yvar' and 'beta_multiplier'
}
\description{
Combine summary statistics
}
