% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sumstats.R
\name{drop_cols_sumstats}
\alias{drop_cols_sumstats}
\title{Drop columns from a sumstats object}
\usage{
drop_cols_sumstats(sumstats, drop_cols)
}
\arguments{
\item{sumstats}{sumstats object}

\item{drop_cols}{column names to drop}
}
\value{
sumstats object without drop_cols
}
\description{
Drop columns from a sumstats object
}
\details{
The intended use of this function is to drop scores from a sumstats object
}
