% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sumstats.R
\name{filter_sumstats}
\alias{filter_sumstats}
\title{Filter a sumstats object on a column in another table}
\usage{
filter_sumstats(sumstats, filter_tbl, name_col, filter_col, filter_threshold)
}
\arguments{
\item{sumstats}{sumstats object}

\item{filter_tbl}{data.frame with filter information}

\item{name_col}{column of filter_tbl with column names to be removed}

\item{filter_col}{column of filter_tbl with values to filter on}

\item{filter_threshold}{remove from sumstats any values of name_col where filter_col < filter_threshold}
}
\value{
sumstats object without columns that fall below the threshold
}
\description{
Filter a sumstats object on a column in another table
}
\details{
The intended use of this function is to filter scores that fall below a threshold of
overlapping variants. name_col should be the score identifier, filter_col should be
the column with values to filter on (overlap or beta_fraction), and filter_threshold
is the minimum allowed value of filter_col.
}
