% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{glmnet_sumstats}
\alias{glmnet_sumstats}
\title{Compute elastic net from summary statistics}
\usage{
glmnet_sumstats(
  xx,
  xy,
  beta_init,
  alpha,
  lambda,
  penalty_factor,
  maxiter = 500,
  tol = 1e-05,
  f.adj = 2,
  verbose = FALSE
)
}
\arguments{
\item{xx}{matrix of x'x where x is nxp design matrix}

\item{xy}{vector of x'y where y is nx1 trait vector}

\item{beta_init}{is px1 vector of starting values}

\item{alpha}{(range 0-1) is fraction of penalty for L1 penalty
and (1-alpha) is fraction for L2 penalty}

\item{lambda}{is penalty parameter}

\item{penalty_factor}{is px1 vector for weighting penalties of design matrix columns.
Typically used as value of 0 for terms not to be penalized and 1 for terms to penalize}

\item{maxiter}{is maximum number of cyclic coordinate descent iterations.
One iteration is over all p parameters.}

\item{tol}{is tolerance to check convergence}

\item{f.adj}{(f.adj >=1) is an adjustment factor to guard against diverging betas, which can
occur when n is small relative to p. The default value of 2 should work most of the time.
If beta parameters are found to diverge, larger values can be used.
Note that larger values of f.adj cause more iterations.
See Yang, Y., & Zou, H. (2012). A coordinate majorization descent algorithm for L1 penalized learning.
Journal of Statistical Computation and Simulation, 84(1), 84–95. https://doi.org/10.1080/00949655.2012.695374}

\item{verbose}{prints summary results at each interation if verbose=TRUE}
}
\value{
list with fitted beta, number of iterations, convegence (true/false), and input penalty parameters
alpha and lambda
}
\description{
Fit an elastic net model based on summary statistics for specified
penalty parameters alpha and lambda
}
\details{
Cyclic coordinate descent is used to fit an elastic net model based on minimizing penalized least squared.
}
\author{
Dan Schaid (schaid@mayo.edu)
}
