% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{glmnet_sumstats_grid}
\alias{glmnet_sumstats_grid}
\title{Fit glmnet over grid of alpha and lambda values from summary statistics}
\usage{
glmnet_sumstats_grid(
  sumstats,
  alpha_grid,
  lambda_frac,
  penalty_factor,
  maxiter = 500,
  tol = 1e-06,
  verbose = FALSE
)
}
\arguments{
\item{sumstats}{list with items xx, xy}

\item{alpha_grid}{vector of alpha values (0-1)}

\item{lambda_frac}{vector of lambda fractions (0-1)}

\item{penalty_factor}{vector of penalty factors}

\item{maxiter}{maximum number of iterations for glmnet_sumstats}

\item{tol}{tolerance for glmnet_sumstats}

\item{verbose}{print progress if TRUE}
}
\value{
matrix (nalpha x nlambda) of glmnet_sumstats results (lists)
}
\description{
Fit glmnet over grid of alpha and lambda values from summary statistics
}
