% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{index_mat_to_vec}
\alias{index_mat_to_vec}
\title{convert matrix indices to vector index}
\usage{
index_mat_to_vec(row_i, col_j, nrow)
}
\arguments{
\item{row_i}{row index}

\item{col_j}{column index}

\item{nrow}{number of rows in matrix}
}
\value{
vector index
}
\description{
convert matrix indices to vector index
}
