% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{loss_indices}
\alias{loss_indices}
\title{loss indices}
\usage{
loss_indices(loss_mean, loss_sd)
}
\arguments{
\item{loss_mean}{mean loss}

\item{loss_sd}{sd of less}
}
\value{
indices
}
\description{
loss indices
}
