% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sumstats_clusters.R
\name{make_sumstats_clusters}
\alias{make_sumstats_clusters}
\title{make summary stats for each cluster
each input is expected to be a data table where the first column is sample ID
this function writes an output file}
\usage{
make_sumstats_clusters(
  trait,
  covariates,
  scores,
  clusters,
  trait_name,
  cohort_name
)
}
\arguments{
\item{trait}{data.frame with two columns: sample ID, outcome}

\item{covariates}{data.frame where first column is sample ID and other columns are numeric covariates}

\item{scores}{data.frame where first column is sample ID and other columns are scores}

\item{clusters}{data.frame with two columns: sample ID, cluster}

\item{trait_name}{name of trait, used to name output file}

\item{cohort_name}{name of cohort, used to name output file}
}
\description{
make summary stats for each cluster
each input is expected to be a data table where the first column is sample ID
this function writes an output file
}
