% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_var_matrix}
\alias{make_var_matrix}
\title{create variance matrix}
\usage{
make_var_matrix(xx, xy, yvar)
}
\arguments{
\item{xx}{X'X matrix}

\item{xy}{X'Y matrix}

\item{yvar}{variance of y}
}
\description{
create variance matrix
}
