% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{make_varxy}
\alias{make_varxy}
\title{create variance matrix}
\usage{
make_varxy(xx, xy, yssq)
}
\arguments{
\item{xx}{X'X matrix}

\item{xy}{X'Y matrix}

\item{yssq}{sum of y^2}
}
\description{
create variance matrix
}
