% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{metrics_sim}
\alias{metrics_sim}
\title{simulate metrics}
\usage{
metrics_sim(wishart_sim, fit_grid, yvar)
}
\arguments{
\item{wishart_sim}{result of \code{\link{sim_sumstats}}}

\item{fit_grid}{2-dimensional grid of fits. Each fit is a list 
and all fits arranged as matrix (of lists)}

\item{yvar}{variance of y}
}
\description{
simulate metrics
}
