% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_sumstats}
\alias{metrics_sumstats}
\title{calculate metrics for summary stats
assumes yvar = 1}
\usage{
metrics_sumstats(sumstats, fit_grid, penalty_factor, trait_type)
}
\arguments{
\item{sumstats}{list with items xx, xy}

\item{fit_grid}{output glmnet_sumstats_grid}

\item{penalty_factor}{used in glmnet_sumstats_grid}

\item{trait_type}{"binary" for case/control, "quant" for quantitative}
}
\value{
list of auc, loss, pen_func,  nbeta, bic_min_index
}
\description{
calculate metrics for summary stats
assumes yvar = 1
}
