% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{r2_glmnet_sumstats}
\alias{r2_glmnet_sumstats}
\title{calculate R2 from glmnet_sumstats results}
\usage{
r2_glmnet_sumstats(beta, xx, xy, eps = 1e-08)
}
\arguments{
\item{beta}{beta from glmnet_sumstats}

\item{xx}{X'X matrix that went into glmnet_sumstats}

\item{xy}{X'y matrix that went into glmnet_sumstats}

\item{eps}{threshold to determine 0}
}
\value{
r2
}
\description{
calculate R2 from glmnet_sumstats results
}
