% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sumstats.R
\name{rename_col_sumstats}
\alias{rename_col_sumstats}
\title{Rename a column in a sumstats object}
\usage{
rename_col_sumstats(sumstats, old_name, new_name)
}
\arguments{
\item{sumstats}{sumstats object}

\item{old_name}{column name to rename}

\item{new_name}{new column name}
}
\value{
sumstats object with renamed column
}
\description{
Rename a column in a sumstats object
}
\details{
The intended use of this function is to rename columns in a sumstats object
}
