% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_sumstats.R
\name{rescale_sumstats}
\alias{rescale_sumstats}
\title{Rescale response-dependent summary statistics}
\usage{
rescale_sumstats(sumstats, scale_multiplier)
}
\arguments{
\item{sumstats}{A summary statistics object containing an `xy` component and
attributes `"ysum"` and `"yssq"`.}

\item{scale_multiplier}{Numeric scalar by which the original response vector
is to be multiplied.}
}
\value{
A summary statistics object of the same structure as `sumstats`,
  with the response-dependent quantities appropriately rescaled.
}
\description{
Rescales the response-dependent components of a summary statistics object
based on response vector \eqn{y}  multiplied by a scalar. If the
transformed response is \eqn{y^* = a y}, then the function updates:
}
\details{
\itemize{
  \item \eqn{X^\top y \rightarrow a X^\top y}
  \item \eqn{\sum y \rightarrow a \sum y}
  \item \eqn{y^\top y \rightarrow a^2 y^\top y}
}

The predictor-dependent quantities (e.g., \eqn{X^\top X}) are left unchanged.
}
\examples{
dat <- sim_test_dat(10, nprs=10)
ss <- make_sumstats(dat$x, dat$y)
ss_scaled <- rescale_sumstats(ss, scale_multiplier = 2)

}
