% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_sumstats}
\alias{sim_sumstats}
\title{simulate y and X matrix from covariance}
\usage{
sim_sumstats(vmat, nsim)
}
\arguments{
\item{vmat}{variance matrix}

\item{nsim}{number of rows}
}
\description{
simulate y and X matrix from covariance
}
