% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{sim_test_dat}
\alias{sim_test_dat}
\title{Simulate example data}
\usage{
sim_test_dat(nsubj, nprs, prev = 0.1, beta.sd = 2)
}
\arguments{
\item{nsubj}{number of subjects}

\item{nprs}{number of PRS}

\item{prev}{prev}

\item{beta.sd}{beta.sd}
}
\description{
Simulate example data
}
