% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_yx}
\alias{simulate_yx}
\title{simulate y and X matrix from covariance
matrix:
       | var(y),  cov(y,x) |
vmat = | cov(x,y) cov(x,x) |
and return matrix [Y|X]}
\usage{
simulate_yx(n, vmat)
}
\arguments{
\item{n}{number of rows}

\item{vmat}{variance matrix}
}
\value{
matrix [Y|X]
}
\description{
simulate y and X matrix from covariance
matrix:
       | var(y),  cov(y,x) |
vmat = | cov(x,y) cov(x,x) |
and return matrix [Y|X]
}
