% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats_to_text.R
\name{sumstats_to_text}
\alias{sumstats_to_text}
\alias{sumstats_from_text}
\title{Convert sumstats to text files}
\usage{
sumstats_to_text(sumstats_file)

sumstats_from_text(sumst_text_files)
}
\arguments{
\item{sumstats_file}{a RDS file ending in .rds that contains a sumstats object}

\item{sumst_text_files}{vector of files produced by sumstats_to_text}
}
\value{
sumstats_to_text: vector of filenames

sumstats_from_text: RDS file containing sumstats object
}
\description{
Convert sumstats to text files and vice versa
}
\details{
sumstats_to_text writes a sumstats object to four text files ending in _xx.txt, _xy.txt,
_colsum.txt and _attr.txt. sumstats_from_text takes these four text files as input
and returns a sumstats object.
}
\examples{
dat <- sim_test_dat(100, 1000, prev=.1, beta.sd=2)
ss1 <- make_sumstats(dat$x, dat$y, center=FALSE)  
ssfile <- paste0(tempfile(), ".rds")
saveRDS(ss1, ssfile)
files <- sumstats_to_text(ssfile)
ssfile2 <- sumstats_from_text(files)
ss2 <- readRDS(ssfile2)
}
