% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{auc_glmnet_sumstats}
\alias{auc_glmnet_sumstats}
\title{calculate AUC from glmnet_sumstats results}
\usage{
auc_glmnet_sumstats(beta, xx, vary, ncase, ncont)
}
\arguments{
\item{beta}{beta from glmnet_sumstats}

\item{xx}{X'X matrix that went into glmnet_sumstats}

\item{vary}{variance of Y}

\item{ncase}{number of cases}

\item{ncont}{number of controls}
}
\value{
list with AUC and R2
}
\description{
calculate AUC from glmnet_sumstats results
}
