% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sumstats.R
\name{combine_sumstats}
\alias{combine_sumstats}
\title{Combine summary statistics}
\usage{
combine_sumstats(sumstats)
}
\arguments{
\item{sumstats}{list of sumstats objects}
}
\value{
list of 1. sumstats object, 2. yvar (which should be 1),
3. beta_multiplier, 4. list of columns with incomplete data (missing in at least
one list element)
}
\description{
Combines summary statistics, centers and scales
}
\details{
Summary statistics in the input list are summed. Columns that are present
in one list element but not others are set to zero when computing the sum.
The results are centered and scaled and divided by the number of observations.
If the sumstats objects provided are not already centered, centering is done
in this function.

Use this function with the results of \code{\link{make_sumstats}}.
}
