% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sumstats.R
\name{combine_sumstats}
\alias{combine_sumstats}
\title{Combine summary statistics}
\usage{
combine_sumstats(sumstats)
}
\arguments{
\item{sumstats}{list with elements xx and xy}
}
\description{
Combine summary statistics
}
