% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sumstats_clusters.R
\name{make_sumstats_clusters}
\alias{make_sumstats_clusters}
\title{Make summary stats for each cluster}
\usage{
make_sumstats_clusters(
  trait,
  covariates,
  scores,
  clusters,
  analysis_name,
  cohort_name,
  min_cluster_size = 20
)
}
\arguments{
\item{trait}{data.frame with two columns: sample ID, outcome}

\item{covariates}{data.frame where first column is sample ID and other columns are numeric covariates}

\item{scores}{data.frame where first column is sample ID and other columns are scores}

\item{clusters}{data.frame with two columns: sample ID, cluster}

\item{analysis_name}{name of analysis, used to name output file}

\item{cohort_name}{name of cohort, used to name output file}

\item{min_cluster_size}{minimum size for a cluster to have an output file}
}
\description{
Make summary stats for each cluster
}
\details{
Each input is expected to be a data table where the first column is sample ID.
Scores with all missing values will be dropped with a message.
Covariates with all missing values will be dropped with a warning.
After applying the above filters, rows with any missing values will be dropped before
calculating the summary statistics.

this function writes an output file with a sumstats object for:
1. all samples together
2. each cluster with size > min_cluster_size
}
