% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{make_var_matrix}
\alias{make_var_matrix}
\title{create variance matrix}
\usage{
make_var_matrix(sumstats)
}
\arguments{
\item{sumstats}{list with items xx, xy, and vary}
}
\description{
create variance matrix
}
