% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIMED_summary_stats.R
\name{metrics_sumstats}
\alias{metrics_sumstats}
\title{calculate metrics for summary stats
assumes yvar = 1}
\usage{
metrics_sumstats(sumstats, fit_grid)
}
\arguments{
\item{sumstats}{list with items xx, xy}

\item{fit_grid}{output of glmnet_sumstats}
}
\value{
list of auc, loss, nbeta
}
\description{
calculate metrics for summary stats
assumes yvar = 1
}
